Renaming components names
