+++ src/screens/login/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Button from '@material-ui/core/Button';
 import Loader from '../../components/loader/Loader';
-import TextInput from '../../components/textInput';
+import TextInput from '../../components/textInput/TextInput';
 
 import Xhr from '../../Xhr';
 import apiEndpoints from '../../apiEndpoints';
